+++ src/lib/connection/XMPPServiceHandler/favoriteEventHandler.ts
@@ -191,8 +191,12 @@ class FavoriteEventHandler extends GenericHandler {
                     case "userpassword":
                         // treated in conversationEventHandler
                         break;
+                    case "conferencerecordingstatus":
+                        // treated in conversationEventHandler
+                        break;
                     case "customStatus":
-                    // treated in presentationEventHandler
+                        // treated in presentationEventHandler
+                        break;
                     default:
                         that._logger.log(that.ERROR, LOG_ID + "(onManagementMessageReceived) unmanaged management message node " + node.getName());
                         break;